refactor(engine): S2 Stage 3 — extract GPU/Diag/HwEnc/Preview-visibility to XlethSvcGpuDiag TU
Verbatim move of the 7 GPU/Diag/HwEnc/Preview-visibility handlers
(Gpu_GetAvailableGpus, Diag_GetVisualPreviewDiagnostic, Gpu_SetAdapter,
HwEnc_GetAvailableEncoders, HwEnc_GetDefaultEncoder, HwEnc_Refresh,
Preview_SetEnabled) out of XlethEngineService.cpp into a new TU, plus the
3 static diag text helpers (readbackHRESULTText, featureLevelText,
readbackFailureStageText) relocated from the file's top anonymous
namespace — all internal-linkage with their sole caller inside the
moved block, so they had to move with it. Zero behavior change.

ctest 43/47, identical to the documented baseline (4 known pre-existing
failures, zero new regressions). Composite diagnostic
(diag_getVisualPreviewDiagnostic) exercised directly against the built
addon on real GPU hardware, confirming all 3 relocated helpers produce
correct output post-move.

See docs/S2_SPLIT_PLAN.md §4 Stage 3.
